+++ src/main/java/net/soulsweaponry/items/sword/GuinsoosRageblade.java
@@ -14,7 +14,7 @@ import net.soulsweaponry.util.TooltipAbilities;
 public class GuinsoosRageblade extends ModdedSword {
 
     public GuinsoosRageblade(ToolMaterial toolMaterial, Settings settings) {
-        super(toolMaterial, ConfigConstructor.rageblade_damage, ConfigConstructor.rageblade_attack_speed, settings);
+        super(toolMaterial, (int) ConfigConstructor.rageblade_damage, ConfigConstructor.rageblade_attack_speed, settings);
         this.addTooltipAbility(TooltipAbilities.FURY, TooltipAbilities.HASTE, TooltipAbilities.FLAME_ENRAGED);
     }
 
@@ -25,7 +25,7 @@ public class GuinsoosRageblade extends ModdedSword {
         }
         if (attacker.isOnFire()) {
             attacker.addStatusEffect(new StatusEffectInstance(StatusEffects.STRENGTH, 200, 2));
-        }    
+        }
         int speed = EnchantmentHelper.getLevel(Enchantments.SWEEPING, stack);
         if (attacker.hasStatusEffect(StatusEffects.HASTE)) {
             StatusEffectInstance effect = attacker.getStatusEffect(StatusEffects.HASTE);
